+++ src/store/media-details/reducer.ts
@@ -10,7 +10,7 @@ export const initialState : MediaDetailsState = {
 const reducer: Reducer<MediaDetailsState> = (state: MediaDetailsState = initialState, action) => {
     switch ((action as MediaDetailsActions).type) {
         case MediaDetailsActionTypes.LOOKUP_TV_SERIES_SUCCESS: {
-            return { ...state, results: action.payload.results, error: "", loading: false };
+            return { ...state, tvlookupresults: action.payload.results, error: "", loading: false };
         }
         case MediaDetailsActionTypes.LOOKUP_TV_SERIES_FAILED: {
             return { ...state, error: action.payload.reason, loading: false };
